Add storybook for the template page view
